# src/main/resources/openapi/api-docs.yml
openapi: 3.0.0
info:
  title: Reviev platform - OpenAPI 1.0
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/platform
...
paths:
  /api/v1/user:
    post:
      operationId: createUser
      summary: Create new user
      tags:
        - User
      parameters:
        - in: header
          name: authorization
          required: true
          schema:
            type: string
      requestBody:
        description: Short user DTO
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShortUserDto'
        required: true
      responses:
        '201':
          description: Created user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '400':
          description: Invalid request
        '404':
          description: Resource not found
        '405':
          description: Method is not available
        '500':
          description: Server error

    put:
      operationId: updateUser
      summary: Update user
      tags:
        - User
      parameters:
        - in: header
          name: authorization
          required: true
          schema:
            type: string
      requestBody:
        description: User DTO
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDto'
        required: true
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '400':
          description: Invalid request
        '404':
          description: Resource not found
        '405':
          description: Method is not available
        '500':
          description: Server error
  /api/v1/users/page={page}&size={size}:
    get:
      tags:
        - User
      summary: Find all users
      operationId: findAllUsers
      parameters:
        - in: header
          name: authorization
          required: true
          schema:
            type: string
        - name: page
          in: path
          required: true
          schema:
            type: integer
            nullable: false
        - name: size
          in: path
          required: true
          schema:
            type: integer
            nullable: false
      responses:
        '200':
          description: Get all users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListDto'
        '400':
          description: Invalid request
        '404':
          description: Resource not found
        '405':
          description: Method is not available
        '500':
          description: Server error
  /api/v1/user/{userId}:
    get:
      tags:
        - User
      summary: Find users by id
      operationId: findUsersById
      parameters:
        - in: header
          name: authorization
          required: true
          schema:
            type: string
        - in: path
          name: userId
          required: true
          schema:
            type: string
            format: uuid
            nullable: false
      responses:
        '200':
          description: Get user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '400':
          description: Invalid request
        '404':
          description: Resource not found
        '405':
          description: Method is not available
        '500':
          description: Server error
    delete:
      tags:
        - User
      summary: Delete user
      operationId: deleteUser
      parameters:
        - in: header
          name: authorization
          required: true
          schema:
            type: string
        - name: userId
          in: path
          description: User id to delete
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Delete user
        '400':
          description: Invalid request
        '404':
          description: Resource not found
        '405':
          description: Method is not available
        '500':
          description: Server error
...
  /api/v1/review:
    post:
      operationId: createReview
      summary: Create new interview
      tags:
        - Review
      parameters:
        - in: header
          name: authorization
          required: true
          schema:
            type: string
      requestBody:
        description: Review DTO
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShortReviewDto'
        required: true
      responses:
        '201':
          description: Created review
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewDto'
        '400':
          description: Invalid request
        '404':
          description: Resource not found
        '405':
          description: Method is not available
        '500':
          description: Server error
  /api/v1/review-for-user:
    post:
      operationId: createReviewForUser
      summary: Create new interview for users
      tags:
        - Review
      parameters:
        - in: header
          name: authorization
          required: true
          schema:
            type: string
      requestBody:
        description: Review DTO
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShortReviewDto'
        required: true
      responses:
        '201':
          description: Created review
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewDto'
        '400':
          description: Invalid request
        '404':
          description: Resource not found
        '405':
          description: Method is not available
        '500':
          description: Server error
  /api/v1/reviews/page={page}&size={size}:
    get:
      tags:
        - Review
      summary: find all review
      operationId: findAllReviews
      parameters:
        - in: header
          name: authorization
          required: true
          schema:
            type: string
        - name: page
          in: path
          required: true
          schema:
            type: integer
            nullable: false
        - name: size
          in: path
          required: true
          schema:
            type: integer
            nullable: false
      responses:
        '200':
          description: Get all review
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewListDto'
        '400':
          description: Invalid request
        '404':
          description: Resource not found
        '405':
          description: Method is not available
        '500':
          description: Server error

#          Ревью Созданные
  /api/v1/reviews/tobe/page={page}&size={size}:
    get:
      tags:
        - Review
      summary: find tobe review
      operationId: findTobeReviews
      parameters:
        - in: header
          name: authorization
          required: true
          schema:
            type: string
        - name: page
          in: path
          required: true
          schema:
            type: integer
            nullable: false
        - name: size
          in: path
          required: true
          schema:
            type: integer
            nullable: false
      responses:
        '200':
          description: Get all review
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewListDto'
        '400':
          description: Invalid request
        '404':
          description: Resource not found
        '405':
          description: Method is not available
        '500':
          description: Server error

#          Ревью Принятые
  /api/v1/reviews/passed/page={page}&size={size}:
    get:
      tags:
        - Review
      summary: find passed review
      operationId: findPassedReviews
      parameters:
        - in: header
          name: authorization
          required: true
          schema:
            type: string
        - name: page
          in: path
          required: true
          schema:
            type: integer
            nullable: false
        - name: size
          in: path
          required: true
          schema:
            type: integer
            nullable: false
      responses:
        '200':
          description: Get all review
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewListDto'
        '400':
          description: Invalid request
        '404':
          description: Resource not found
        '405':
          description: Method is not available
        '500':
          description: Server error

#          Найти отмененные ревью
  /api/v1/reviews/canceled/page={page}&size={size}:
    get:
      tags:
        - Review
      summary: find canceled review
      operationId: findCanceledReviews
      parameters:
        - in: header
          name: authorization
          required: true
          schema:
            type: string
        - name: page
          in: path
          required: true
          schema:
            type: integer
            nullable: false
        - name: size
          in: path
          required: true
          schema:
            type: integer
            nullable: false
      responses:
        '200':
          description: Get all review
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewListDto'
        '400':
          description: Invalid request
        '404':
          description: Resource not found
        '405':
          description: Method is not available
        '500':
          description: Server error
  /api/v1/review/revievId={reviewId}&isDone={isDone}:
    post:
      tags:
        - Review
      summary: close review
      operationId: closeReview
      parameters:
        - in: header
          name: authorization
          required: true
          schema:
            type: string
        - name: reviewId
          in: path
          required: true
          schema:
            type: string
            format: uuid
            nullable: false
        - name: isDone
          in: path
          required: true
          schema:
            type: boolean
            nullable: false
      requestBody:
        description: Authentication request
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthenticationToken'
        required: true
      responses:
        '200':
          description: Review close
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewDto'
        '400':
          description: Invalid request
        '404':
          description: Resource not found
        '405':
          description: Method is not available
        '500':
          description: Server error
  /api/v1/review/{reviewId}:
    get:
      tags:
        - Review
      summary: Find review by id
      operationId: findReviewById
      parameters:
        - in: header
          name: authorization
          required: true
          schema:
            type: string
        - name: reviewId
          in: path
          required: true
          schema:
            type: string
            format: uuid
            nullable: false
      responses:
        '200':
          description: Get review
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewDto'
        '400':
          description: Invalid request
        '404':
          description: Resource not found
        '405':
          description: Method is not available
        '500':
          description: Server error
    delete:
      tags:
        - Review
      summary: Delete review
      operationId: deleteReview
      parameters:
        - name: reviewId
          in: path
          description: Review id to delete
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Delete review
        '400':
          description: Invalid request
        '404':
          description: Resource not found
        '405':
          description: Method is not available
        '500':
          description: Server error
    put:
      operationId: updateReview
      summary: Update review
      tags:
        - Review
      parameters:
        - in: header
          name: authorization
          required: true
          schema:
            type: string
        - name: reviewId
          in: path
          description: Review by id to update
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        description: Review DTO
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewDto'
        required: true
      responses:
        '200':
          description: Review updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewDto'
        '400':
          description: Invalid request
        '404':
          description: Resource not found
        '405':
          description: Method is not available
        '500':
          description: Server error


  # --- Commentary ---
...
components:
  schemas:
...
    UserDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: User id
        firstName:
          type: string
          description: User first name
        lastName:
          type: string
          description: User last name
        avatar:
          type: string
          description: User avatar url
        email:
          type: string
          description: User email
        role:
          type: string
          enum:
            - USER
            - ADMIN
        specialities:
          type: string
          enum:
            - FRONTEND
            - BACKEND
        reviewsReceiving:
          type: array
          items:
            $ref: '#/components/schemas/ReviewDto'
        reviewsStudent:
          type: array
          items:
            $ref: '#/components/schemas/ReviewDto'
        socials:
          type: array
          items:
            $ref: '#/components/schemas/SocialDto'
    UserListDto:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/UserDto'
        totalElements:
          type: number
        totalPages:
          type: number
        currentPage:
          type: number
        last:
          type: boolean
        numberOfElements:
          type: number
        first:
          type: boolean
        empty:
          type: boolean
    ShortUserDto:
      type: object
      properties:
        firstName:
          type: string
          description: User first name
        lastName:
          type: string
          description: User last name
        password:
          type: string
          description: User password
        email:
          type: string
          description: User email
    SocialDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Social id
        name:
          type: string
          description: Social name
        link:
          type: string
          description: Social link
...
    ReviewDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        theme:
          type: string
        time:
          type: string
          format: date-time
        link:
          type: string
        speciality:
          type: string
          enum:
            - FRONTEND
            - BACKEND
        status:
          type: string
          enum:
            - TOBE
            - PASSED
            - CANCELED
        reviewer:
          $ref: '#/components/schemas/UserDto'
        student:
          $ref: '#/components/schemas/UserDto'
    ShortReviewDto:
      type: object
      properties:
        theme:
          type: string
        time:
          type: string
          format: date-time
        link:
          type: string
        speciality:
          type: string
          enum:
            - FRONTEND
            - BACKEND
        authData:
          $ref: '#/components/schemas/AuthenticationToken'
    ReviewListDto:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ReviewDto'
        totalElements:
          type: number
        totalPages:
          type: number
        currentPage:
          type: number
        last:
          type: boolean
        numberOfElements:
          type: number
        first:
          type: boolean
        empty:
          type: boolean
...
    AuthenticationToken:
      type: object
      properties:
        token:
          type: string
          description: JWT